Rename StateDistribution files
