+++ lineage/tests/test_LineageTree.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 from ..CellVar import CellVar as c, get_subtrees, find_two_subtrees, get_mixed_subtrees
 from ..LineageTree import LineageTree, max_gen, get_leaves
-from ..states.StateDistribution import StateDistribution
+from ..states.StateDistributionGamma import StateDistribution
 from ..states.stateCommon import get_experiment_time
